T5: extract star-review labeling UI into reusable plugin control
Move the interactive star-review labeling UI out of TestApp and into the
plugin (NINA.Joko.Plugins.HocusFocus.StarDetection.Optimization.Review) so
both TestApp's `review` dev tool and the future in-NINA wizard share one
source of truth. Behavior-preserving refactor — no detector/options/label-JSON
changes.

What moved:
- StarReviewLabels.cs, StarReviewViewport.cs (pure) — straight move + namespace.
- StarReviewVM.cs (+ FrameReview, marker models, LabelPass) — decoupled from
  TestApp via a per-frame Func<Task<BitmapSource>> ImageProvider seam; dropped
  the IProfileService dep and the AttachWindow(Window) coupling.
- Window visual tree + mouse/keyboard handlers → new StarReviewControl
  UserControl (binds to StarReviewVM as DataContext).
- New StarReviewImaging plugin helper owns the MTF stretch + Mat→BitmapSource;
  TestApp's Program.ToBitmapSource now delegates to it.

TestApp now hosts the plugin control inside a thin StarReviewWindow and wires
Closing→vm.SaveAll() in the STA runner. Tests .csproj drops the Labels/Viewport
compile-links (now via the plugin ProjectReference); StarReviewTests usings
updated. Build clean, 1127 tests pass.
