+++ Joko.NINA.Plugins/TestApp/DiagnoseLabelsRunner.cs
@@ -13,6 +13,7 @@
 using NINA.Core.Enum;
 using NINA.Joko.Plugins.HocusFocus.Interfaces;
 using NINA.Joko.Plugins.HocusFocus.StarDetection;
+using NINA.Joko.Plugins.HocusFocus.StarDetection.Optimization.Review;
 using NINA.Joko.Plugins.HocusFocus.Utility;
 using NINA.Profile;
 using NINA.Profile.Interfaces;
